requirements.txt
web3>=6.0.0
requests>=2.28.0
python-dotenv>=1.0.0
prometheus-client>=0.17.0
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
aiohttp>=3.8.0
structlog>=23.1.0
python-json-logger>=2.0.7
slack-sdk>=3.19.0
python-telegram-bot>=20.0